+++ src/main/java/com/rnpc/inventory/repository/CellphonePartsRepository.java
@@ -3,5 +3,5 @@ package com.rnpc.inventory.repository;
 import com.rnpc.inventory.entity.CellphoneParts;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CellphonePartsRepository extends JpaRepository<CellphoneParts, Integer> {
+public interface CellphonePartsRepository extends JpaRepository<CellphoneParts, Long> {
 }
